some code refactor
